+++ apps/docs-website/src/components/RulesTable.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unnecessary-condition */
 import React, { useEffect, useRef, useState } from "react";
 import {
   createColumnHelper,
@@ -6,18 +7,15 @@ import {
   getFilteredRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-import {
-  pluginsNamesWIthJest,
-  pluginsNamesWIthVitest,
-  rulesetWithJest,
-  rulesetWithVitest,
-} from "@sheriff/utils";
-import type { Entry } from "@sheriff/types";
+import type { Entry, ServerResponse, SheriffSettings } from "@sheriff/types";
 import { isEmpty } from "lodash-es";
 import Select from "react-select";
 import { DebounceInput } from "react-debounce-input";
 import { filterDuplicateRules } from "../utils/filterDuplicatedRules";
 import styles from "./RulesTable.module.css";
+import { ConfigCombinationForm } from "./ConfigCombinationForm";
+import { TableSkeleton } from "./TableSkeleton";
+import { configCombinationDefaultValues } from "./constants";
 
 const columnHelper = createColumnHelper<Entry>();
 
@@ -53,16 +51,16 @@ const columns = [
   }),
 ];
 
-const dedupedRulesetWithVitest = filterDuplicateRules(rulesetWithVitest);
-const dedupedRulesetWithJest = filterDuplicateRules(rulesetWithJest);
-
 export const RulesTable = (): JSX.Element => {
-  const [data, setData] = useState(() => [...dedupedRulesetWithVitest]);
+  const [data, setData] = useState<Entry[]>(() => []);
+  const [isLoading, setIsLoading] = useState(false);
+  const [pluginsNames, setPluginsNames] = useState<string[]>([]);
+  const [configCombination, setConfigCombination] = useState<SheriffSettings>(
+    configCombinationDefaultValues,
+  );
   const [globalFilter, setGlobalFilter] = React.useState("");
   const [inputValue, setInputValue] = React.useState("");
   const [selectValue, setSelectValue] = React.useState(null);
-  const [isVitestChecked, setIsVitestChecked] = useState<boolean>(true);
-  const [isJestChecked, setIsJestChecked] = useState<boolean>(false);
   const [tableMaximumAllowedWidth, setTableMaximumAllowedWidth] =
     useState<number>(0);
   const tableContainerRef = useRef<HTMLDivElement>(null);
@@ -77,36 +75,45 @@ export const RulesTable = (): JSX.Element => {
     setTableMaximumAllowedWidth(computedTableWidth);
   }, []);
 
+  useEffect(() => {
+    const fetchFreshRuleset = async () => {
+      setIsLoading(true);
+      try {
+        const response = await fetch(
+          process?.env?.NODE_ENV === "development"
+            ? "http://localhost:5000/api/get-new-sheriff-config"
+            : "https://sheriff-webservices.onrender.com/api/get-new-sheriff-config",
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+              "Access-Control-Allow-Origin": "*",
+            },
+            body: JSON.stringify(configCombination),
+          },
+        );
+
+        const fetchedData: ServerResponse = await response.json();
+
+        setPluginsNames(fetchedData.pluginsNames);
+
+        setData(filterDuplicateRules(fetchedData.compiledConfig));
+      } catch (error) {
+        console.error(error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchFreshRuleset().catch(console.error);
+  }, [configCombination]);
+
   const resetSelectValue = () => {
     setSelectValue(null);
   };
 
   const resetInputValue = () => {
     setInputValue("");
-  };
-
-  const handleVitestCheckbox = () => {
-    if (isVitestChecked) {
-      setData([...dedupedRulesetWithJest]);
-      setIsVitestChecked(false);
-      setIsJestChecked(true);
-    } else {
-      setData([...dedupedRulesetWithVitest]);
-      setIsVitestChecked(true);
-      setIsJestChecked(false);
-    }
-  };
-
-  const handleJestCheckbox = () => {
-    if (isJestChecked) {
-      setData([...dedupedRulesetWithVitest]);
-      setIsVitestChecked(true);
-      setIsJestChecked(false);
-    } else {
-      setData([...dedupedRulesetWithJest]);
-      setIsVitestChecked(false);
-      setIsJestChecked(true);
-    }
   };
 
   const table = useReactTable({
@@ -121,57 +128,11 @@ export const RulesTable = (): JSX.Element => {
   });
 
   const cellMaxWidth = tableMaximumAllowedWidth / columns.length;
-  const pluginsNames = isVitestChecked
-    ? pluginsNamesWIthVitest
-    : pluginsNamesWIthJest;
 
   return (
     <div ref={tableContainerRef}>
       <div className={styles.tableControlsContainer}>
-        <div className={styles.checkboxGroupContainer}>
-          <div className={styles.nativeCheckboxDisabled}>
-            <input checked disabled type="checkbox" name="react" id="react" />
-            <label htmlFor="react">React</label>
-          </div>
-          <div className={styles.nativeCheckboxDisabled}>
-            <input checked disabled type="checkbox" name="lodash" id="lodash" />
-            <label htmlFor="lodash">Lodash</label>
-          </div>
-          <div className={styles.nativeCheckboxDisabled}>
-            <input checked disabled type="checkbox" name="next" id="next" />
-            <label htmlFor="next">Next</label>
-          </div>
-          <div className={styles.nativeCheckboxDisabled}>
-            <input
-              checked
-              disabled
-              type="checkbox"
-              name="playwright"
-              id="playwright"
-            />
-            <label htmlFor="playwright">Playwright</label>
-          </div>
-          <div className={styles.nativeCheckbox}>
-            <input
-              type="checkbox"
-              name="vitest"
-              id="vitest"
-              checked={isVitestChecked}
-              onChange={handleVitestCheckbox}
-            />
-            <label htmlFor="vitest">Vitest</label>
-          </div>
-          <div className={styles.nativeCheckbox}>
-            <input
-              type="checkbox"
-              name="jest"
-              id="jest"
-              checked={isJestChecked}
-              onChange={handleJestCheckbox}
-            />
-            <label htmlFor="jest">Jest</label>
-          </div>
-        </div>
+        <ConfigCombinationForm setTableData={setConfigCombination} />
         <div className={styles.filtersContainer}>
           <DebounceInput
             className={styles.filterInput}
@@ -253,47 +214,51 @@ export const RulesTable = (): JSX.Element => {
           />
         </div>
       </div>
-      <table>
-        <thead>
-          {table.getHeaderGroups().map((headerGroup) => (
-            <tr key={headerGroup.id}>
-              {headerGroup.headers.map((header) => (
-                <th
-                  key={header.id}
-                  className={styles.th}
-                  style={{
-                    maxWidth: cellMaxWidth,
-                  }}
-                >
-                  {header.isPlaceholder
-                    ? null
-                    : flexRender(
-                        header.column.columnDef.header,
-                        header.getContext(),
-                      )}
-                </th>
-              ))}
-            </tr>
-          ))}
-        </thead>
-        <tbody>
-          {table.getRowModel().rows.map((row) => (
-            <tr key={row.id} className={styles.tr}>
-              {row.getVisibleCells().map((cell) => (
-                <td
-                  key={cell.id}
-                  className={styles.td}
-                  style={{
-                    maxWidth: cellMaxWidth,
-                  }}
-                >
-                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
+      {isLoading ? (
+        <TableSkeleton />
+      ) : (
+        <table>
+          <thead>
+            {table.getHeaderGroups().map((headerGroup) => (
+              <tr key={headerGroup.id}>
+                {headerGroup.headers.map((header) => (
+                  <th
+                    key={header.id}
+                    className={styles.th}
+                    style={{
+                      maxWidth: cellMaxWidth,
+                    }}
+                  >
+                    {header.isPlaceholder
+                      ? null
+                      : flexRender(
+                          header.column.columnDef.header,
+                          header.getContext(),
+                        )}
+                  </th>
+                ))}
+              </tr>
+            ))}
+          </thead>
+          <tbody>
+            {table.getRowModel().rows.map((row) => (
+              <tr key={row.id} className={styles.tr}>
+                {row.getVisibleCells().map((cell) => (
+                  <td
+                    key={cell.id}
+                    className={styles.td}
+                    style={{
+                      maxWidth: cellMaxWidth,
+                    }}
+                  >
+                    {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                  </td>
+                ))}
+              </tr>
+            ))}
+          </tbody>
+        </table>
+      )}
     </div>
   );
 };