+++ packages/eslint-config-sheriff/src/getExportableConfig.ts
@@ -15,12 +15,8 @@ import vitest from 'eslint-plugin-vitest';
 import preferEarlyReturn from '@regru/eslint-plugin-prefer-early-return';
 import tsdoc from 'eslint-plugin-tsdoc';
 import storybook from 'eslint-plugin-storybook';
-import {
-  allJsExtensions,
-  supportedFileTypes,
-  ignores,
-  allJsxExtensions,
-} from './constants';
+import fsecond from 'eslint-plugin-fsecond';
+import { allJsExtensions, supportedFileTypes, ignores } from './constants';
 import { fpHandPickedRules } from './fpHandPickedRules';
 import { getBaseEslintHandPickedRules } from './getBaseEslintHandPickedRules';
 import { getReactConfig } from './getReactConfig';
@@ -250,6 +246,11 @@ const getBaseConfig = (userConfigChoices: SheriffSettings) => {
           mode: 'typescript',
         },
       },
+    },
+    {
+      files: [supportedFileTypes],
+      plugins: { fsecond },
+      rules: { 'fsecond/prefer-destructured-optionals': 2 },
     },
     getAstroConfig(hasReact, customTSConfigPath),
     {