+++ packages/eslint-config-sheriff/src/getReactConfig.ts
@@ -2,6 +2,7 @@ import react from 'eslint-plugin-react';
 import reactAccessibility from 'eslint-plugin-jsx-a11y';
 import reactHooks from 'eslint-plugin-react-hooks';
 import reactRefresh from 'eslint-plugin-react-refresh';
+import fsecond from 'eslint-plugin-fsecond';
 import {
   allJsExtensions,
   allJsxExtensions,
@@ -55,11 +56,20 @@ export const getReactConfig = (customTSConfigPath?: string | string[]) => {
       rules: reactAccessibility.configs.recommended.rules,
     },
     {
-      files: [`**/*{${allJsExtensions},${allJsxExtensions}}`],
+      files: [`**/*{${allJsxExtensions}}`],
       plugins: {
         'react-hooks': reactHooks,
       },
       rules: reactHooks.configs.recommended.rules,
+    },
+    {
+      files: [`**/*{${allJsxExtensions}}`],
+      plugins: {
+        fsecond,
+      },
+      rules: {
+        'fsecond/valid-event-listener': 2,
+      },
     },
     // Specific overrides for storybook
     {